requirements.txt
speedtest-cli
ping3
paho-mqtt